+++ src/constants/main.js
@@ -4,6 +4,8 @@ export const FILM_CARD_ACTIVATION_DELAY = 1000;
 
 export const MINUTES_IN_HOUR = 60;
 
+export const EXTRA_FIMS_COUNT = 4;
+
 export const MONTHS = [`January`, `February`, `March`, `April`, `May`, `June`, `July`, `August`, `September`, `October`, `November`, `December`];
 
 export const PromoSetting = {
